Add a go.mod file.
fileutil.ReadDir got a new interface, and rather than update to use it, decided to embed the func instead since it's so trivial.  Then I started looking to see if we could get rid of the fileutil dependency.  In the end there were a few calls that were impractical to get rid of or embed, but I did remove the Fdatasync calls: from what I can tell this was only valuable pre Go 1.12 for MacOS, whereas now the builtin os.File.Sync does the right thing.